Modified get_param_value_from_conf_file() function to throw an Exception.
get_param_value_from_conf_file() function used to print a message and
finish the program with sys.exit(1). That behaviour is very difficult to
debug, and was causing errors in the sphinx autodoc module, due to unset
CONF_INI_FILE_PATH environment variable. That environment variable is
required in order to generate the documentation properly.